Remove streaming foyer pages
Sarah asked that the Streaming foyer page be removed from this initial release of runtime.
Correspondingly, create (temporary?) Operations foyer page to contain the Oozie publication orphaned by the removal of the Streaming foyer page.
In a separate meeting, there was concern about having foyer pages that match the name of the traditional stack groupings:
"core", "data warehouse", "data science", "operational DB", "secuity", "governance", "streaming" and "operations" --
as the thinking was these were too "component based" when we wanted the organization to be more use-case driven.
Given that we have precident for including "secuity", "governance", "streaming" as foyer pages,
I've (temporarily) created an operations foyer page and added the Oozie publication ("Configuring Oozie") to it.
Email sent to Sairam describing the situation. He can endorse the current structure or suggest a new one.