Reverted changes.
If you're trying to re-setup the foyer page at this location,
please look at the configuration of another library that
has foyer pages for a model of how it's done.
For example, the href should be up one level,
and you still need to include an entry for each publication
so we can provide a complete left-nav TOC for the subtopics.